Fixed some compiler warnings (#2426)
* Added warnings to examples.

* fixed some warnings

* using button sender parameter.

* use button sender in example

* use button sender in example

* removed sender from methods that don’t use it.

* removed sender from methods that don’t use it.

* use button sender in example

* use button sender in example

* use button sender in example

* use button sender in example

* Fixed init of header configurator to use passed in parameter

* removed unused API

* use button sender in example

* Revert "Added warnings to examples."

This reverts commit 91f0480f86cb609c47b00f1346f82de67c31c4e9.

* fixed example: Using the view controllers navigationItem rather than creating a new one.

* removed finished checks from animation blocks in examples.

* reverted id sender checks

* revert id sender check for Flexible header UINavigationBar

* revert clug

* revert clug

* addressing nits

* removed sender check for FeatureHighlightTypicalUseView

* more nits

+++ components/Buttons/examples/supplemental/ButtonsTypicalUseSupplemental.m
@@ -82,7 +82,7 @@
   CGFloat lastButtonMaxY = (CGRectGetHeight(lastButton.bounds) / 2) + lastButton.center.y;
   if (lastButtonMaxY > CGRectGetHeight(self.view.bounds)) {
     CGFloat columnOffset = CGRectGetWidth(self.view.bounds) / 4;
-    NSUInteger splitIndex = ceil(self.buttons.count / 2);
+    NSUInteger splitIndex = (NSUInteger)ceil(self.buttons.count / 2);
 
     if (((MDCButton *)self.buttons[splitIndex - 1]).enabled) {
       splitIndex++;
@@ -116,7 +116,7 @@
   MDCButton *lastButton = self.buttons[NSMaxRange(range) - 1];
   CGFloat verticalCenterY =
       (viewHeight - (lastButton.center.y + (CGRectGetHeight(lastButton.bounds) / 2))) / 2;
-  for (NSInteger i = range.location; i < NSMaxRange(range); i++) {
+  for (NSUInteger i = range.location; i < NSMaxRange(range); i++) {
     MDCButton *button = self.buttons[i];
     UILabel *label = self.labels[i];
     button.center = MDCRoundCenterWithBoundsAndScale(CGPointMake(button.center.x,
